% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_log_probs.R
\name{normalize_log_probs}
\alias{normalize_log_probs}
\title{A function that calculates the normalized probabilities given the unormalized log probabilities}
\usage{
normalize_log_probs(
  log_probs,
  na.rm = FALSE,
  verbose = FALSE,
  threshhold = 10^(-2)
)
}
\arguments{
\item{log_probs}{a numeric vector containing unormalized log probabilities}

\item{na.rm}{a boolean indicating whether to remove NA values}

\item{verbose}{a boolean indicating whether to print function progress}

\item{threshhold}{a float that is a threshold for numeric error}
}
\value{
a numeric vector of normalized probabilities
}
\description{
A function that calculates the normalized probabilities given the unormalized log probabilities
}
\examples{
normalize_log_probs(c(-10, -10, -10))
normalize_log_probs(c(NA, -10, -10, -10), na.rm = TRUE)
}
